% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{add}
\alias{add}
\title{A function for adding}
\usage{
add(x, y)
}
\arguments{
\item{x}{a number}

\item{y}{another number}
}
\value{
a number
}
\description{
A really good adding function.
Perhaps the best function ever!

A work of pure genius.
}
\examples{
add(5, 10)
## Can also use negative numbers
add(-5, 10)
}
